@@ -4,7 +4,7 @@
   "description": "Ireneus: Blog",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon --use-strict index.js",
+    "start": "nodemon --use-strict index.js && xdg-open http://ireneus:3002/ ",
     "bundle": "webpack",
     "lint": "eslint webpack.config.js  server/routes"
   },
@@ -20,7 +20,7 @@
     "express": "^4.14.0",
     "jsonwebtoken": "^7.4.1",
     "material-ui": "^0.16.5",
-    "mongoose": "^4.10.8",
+    "mongoose": "^4.11.3",
     "passport": "^0.3.2",
     "passport-local": "^1.0.0",
     "react": "^15.4.1",
@@ -34,9 +34,11 @@
     "babel-loader": "^6.2.10",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
-    "eslint": "^4.1.1",
-    "eslint-loader": "^1.8.0",
+    "chai": "^4.1.0",
+    "eslint": "^4.2.0",
+    "eslint-loader": "^1.9.0",
     "eslint-plugin-react": "^7.1.0",
+    "mocha": "^3.4.2",
     "nodemon": "^1.11.0",
     "webpack": "^1.14.0"
   }
