@@ -1,98 +1,121 @@
-import React from 'react' 
-import LoginForm from '../components/LoginForm.jsx' 
+import React, { PropTypes } from 'react';
 import Auth from '../modules/Auth';
+import LoginForm from '../components/LoginForm.jsx';
 
-class LoginPage extends React.Component{
-    constructor(props){
-        super(props) 
-        
-        const storedMessage =localStorage.getItem('successMessage') 
-        let successMessage = ''
 
-        if(storedMessage){
-            successMessage = storedMessage
-            localStorage.removeItem('successMessage')
+class LoginPage extends React.Component {
+
+    /**
+     * Class constructor.
+     */
+    constructor(props, context) {
+        super(props, context);
+
+        const storedMessage = localStorage.getItem('successMessage');
+        let successMessage = '';
+
+        if (storedMessage) {
+            successMessage = storedMessage;
+            localStorage.removeItem('successMessage');
         }
+
+        // set the initial component state
         this.state = {
-            errors : {},
-            user : {
-                email : '',
-                password : ''
+            errors: {},
+            successMessage,
+            user: {
+                email: '',
+                password: ''
             }
-        }
+        };
+
+        this.processForm = this.processForm.bind(this);
+        this.changeUser = this.changeUser.bind(this);
     }
 
-    processForm(event){
-        event.preventDefault() 
+    /**
+     * Process the form.
+     *
+     * @param {object} event - the JavaScript event object
+     */
+    processForm(event) {
+        // prevent default action. in this case, action is the form submission event
+        event.preventDefault();
 
         // create a string for an HTTP body message
-        const email = encodeURIComponent(this.state.user.email) 
-        const password = encodeURIComponent(this.state.user.password) 
-        const formData = `email=${email}&password=${password}` 
+        const email = encodeURIComponent(this.state.user.email);
+        const password = encodeURIComponent(this.state.user.password);
+        const formData = `email=${email}&password=${password}`;
 
         // create an AJAX request
-        const xhr = new XMLHttpRequest() 
-        xhr.open('post', '/auth/login') 
-        xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded') 
-        xhr.responseType = 'json' 
+        const xhr = new XMLHttpRequest();
+        xhr.open('post', '/auth/login');
+        xhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
+        xhr.responseType = 'json';
         xhr.addEventListener('load', () => {
             if (xhr.status === 200) {
                 // success
-                Auth.authenticateUser(xhr.response.token)
+
                 // change the component-container state
                 this.setState({
                     errors: {}
-                }) 
+                });
 
-                console.log('The form is valid')
+                // save the token
+                Auth.authenticateUser(xhr.response.token);
 
-                //Change current Url to /
-                this.context.route.replace('/')
+
+                // change the current URL to /
+                this.context.router.replace('/');
             } else {
                 // failure
 
                 // change the component state
-                const errors = xhr.response.errors ? xhr.response.errors : {} 
-                errors.summary = xhr.response.message 
+                const errors = xhr.response.errors ? xhr.response.errors : {};
+                errors.summary = xhr.response.message;
 
                 this.setState({
                     errors
-                }) 
+                });
             }
-        }) 
-        xhr.send(formData) 
-
-
-        console.log('email:', this.state.user.email) 
-        console.log('password:', this.state.user.password) 
+        });
+        xhr.send(formData);
     }
 
-    changeUser(event){
-        const field = event.target.name 
-        const user = this.state.user 
-        user[field] = event.target.value 
+    /**
+     * Change the user object.
+     *
+     * @param {object} event - the JavaScript event object
+     */
+    changeUser(event) {
+        const field = event.target.name;
+        const user = this.state.user;
+        user[field] = event.target.value;
 
         this.setState({
             user
-        })
+        });
     }
 
-    render(){
-        return(
+    /**
+     * Render the component.
+     */
+    render() {
+        return (
             <LoginForm
                 onSubmit={this.processForm}
                 onChange={this.changeUser}
                 errors={this.state.errors}
+                successMessage={this.state.successMessage}
                 user={this.state.user}
             />
-        )
+        );
     }
 
 }
 
-LoginPage.PropTypes = {
-    errors: React.PropTypes.object.isRequired,
-    user : React.PropTypes.object.isRequired
-} 
+LoginPage.contextTypes = {
+    router: PropTypes.object.isRequired
+};
 
-export default LoginPage 
+export default LoginPage;
